Use invokable classes for handlers
